added pmd-conform file-layer
git-svn-id: https://svn.uni-konstanz.de/disy/Treetank/trunk/treetank@4864 e3ddb328-5bfe-0310-b762-aafcbcbd2528

+++ src/main/java/com/treetank/utils/ICrypto.java
@@ -22,9 +22,8 @@ import com.treetank.io.file.ByteBufferSinkAndSource;
 
 public interface ICrypto {
 
-	public short crypt(final short length, final ByteBufferSinkAndSource buffer);
+	public int crypt(final int length, final ByteBufferSinkAndSource buffer);
 
-	public short decrypt(final short length,
-			final ByteBufferSinkAndSource buffer);
+	public int decrypt(final int length, final ByteBufferSinkAndSource buffer);
 
 }
